Change handling of tokens
Using file inputs for the token is not only annoying but also unlikely
to be helpful, because when running as a GitHub action, we'll want to
use environment variables anyway. For testing, it's convenient to pass
the tokens on the command line, so that's what it now does.

Also changed the names of some of the options for greater clarity.