If have no CM & CFF file, act like --all-metadata is on
In the case where a repo has no CM or CFF file, the resulting record
is unexpected sparse unless --all-metadata is turned on. It seems to
me to make more sense to always act as if --all-metadata has been
given for repos that lack both codemeta and citation.cff files.