Change the option --all-metadata to --leaner-metadata
If we want the default behavior to be roughly the same whether the
user has a codemeta or citation.cff file or not, then the default has
to what "all-metadata" did. So the sense has to be reversed and we
need an option to reduce how much metadata is included.