fix: don't put codemeta "provider" in contributors
I'm not quite clear on whether provider can be a person that can be
counted as a contributor. In at least one example repo, someone set
"provider" to CRAN, which makes sense as a value, and that in turn
implies maybe it doesn't make sense to put "provider" in a record's
"contributors" list.