fix: don't try to use award id's after all
InvenioRDM checks the values (somehow), but at the same time, we can't
be sure what _type_ of identifiers the user is providing, so we're
very likely to cause an error if we just put the award id in there. So
I guess all this effort to get the funding info has been a waste of
timeo.