fix: spacy loading & parsing issues
SpaCy models weren't being installed by requirements.txt, which meant
name parsing would fail if it had to resort to spaCy.

Also, spaCy's latest transformer models behave differently with
respect to classifying human names (they seem to be better), so I had
to adjust the algorithm in is_person() to account for that.