feat: download spaCy models automatically if needed
IGA will now download needed spaCy models itself in a one-time step at
run-time. This is necessary because putting the download into the
`requirements.txt` file makes it impossible to put IGA on PyPI, due to
PyPI disallowing direct requirements. The change to IGA means that we
take a huge run-time performance hit if spaCy is needed, but at least
it'll happen only once and not again on subsequent runs of IGA.