fix: print errors only once
The previous arrangement produced this in the output:
```
Run caltechlibrary/iga@main
Run echo "error=" >> "$GITHUB_ENV"
Error: Value of INVENIO_TOKEN is not set.
Run actions/github-script@v6
Error: Workflow stopped due to an error.
```

But it seems that GitHub only prints the last error, so what you would
see in the summary and in the email message is `Error: Workflow
stopped due to an error.`, which was not helpful.