fix: redirected output in normal mode didn't get redirected
This endless futzing with redirection and log destinations and so on
turn out to be due to the fact that the Click library processes eager
options in the order they appear on the command line, and doesn't have
(or I haven't found) a way to tell it to always process certain
options first.

The current situation is still not ideal, but at least there's a way
to get output even in normal mode. The trick is to put the
`--log-dest` option first on the command line:

```
iga --log-dest file.out --mode verbose ...
```