fix: issue #8
The original problem was that `_org_from_dict(...)` was used for both
organizational entities used as (e.g.) creators or contributors, and
for affiliations. It turns out that the InvenioRDM requirements for
the dict in these two cases is different: one uses a field named
`identifier` and the other names it `id`.

While working on the solution, I realized that the algorithm could be
improved. There's no good reason not to use the value of a `name`
field if it's provided, so we try that first instead of trying to get
it from the identifier. Also, the function now tries to look up org
names in ROR if it's given an identifier and no name.