fix: issue #21 -- bogus copyright date in some cases
A `codemeta.json` file can include a value for `copyrightYear` and
that value can be an integer and not a string. In the function
`dates()` in `metadata.py`, the value gets handed to the Python Arrow
date parser package as-is. Apparently you get unexpected results when
you hand it an integer and not string. The fix is simple.