feat: revise the sample workflow & edit text a bit
This now gets the host name of the inveniordm server and puts that
name into the step name as "Send to servername" instead of the
previous step name, "Send to InvenioRDM". The motivation is that users
of repositories may not even know what "InvenioRDM" is, because the
branding of a given institution's repository may not mention it. So,
users seeing the workflow may not find "Send to InvenioRDM" very
helpful.

Also, the name "GitHub Action" should be "GitHub Actions" (plural), so
I edited the text to adjust things appropriately.