fix: need a version of the name to use in code
A problem with using the "name" field from codemeta.json for the full
name of the program is that the result is not usable in (e.g.)
Makefiles. After looking at the CodeMeta spec, which doesn't clearly
spell out what to do, and examples of what people are doing in
codemeta.json files on GitHub, I'm going to start using the
CodeMeta "identifier" field for the program name. This is what other
people seem to do, when they use the "identifier" field value in their
codemeta.json files.