fix: install IGA from the develop branch when archiving IGA
We need a custom version of the IGA workflow so that we can do
releases normally and have the release use the latest version of IGA
instead of the version on PyPI. The latter version will be one behind
the release at the time a release on GitHub is made, because we don't
update PyPI until we made the GitHub release and have sent it to our
RDM sever to get a DOI that we write back to the README.md and
codemeta.json files.

The approach taken here, of using the version from the develop branch,
makes an assumption that the develop branch hast the latest code at
the time of the release on GitHub. This should always be the case
because we work off the develop branch and merge to main when we do
releases. But there's a risk that the main branch has newer commits
than the develop branch. I don't currently have a solution to that.