fix: refactor to address issue #14
This now attempts to make is_person() handle names written in Chinese,
Japanese and Korean scripts. It uses multiple heuristics to do this,
and it is not very accurate right now. However, it's still better than
it was before, because before, is_person() would simply return False
for name written in CJK scripts.