fix: handle missing publisher fields in invenio_server_name()
`invenio_server_name(...)` tries to get the name of the server by
getting the first record and looking in the field metadata.publisher
in that record. It turns out our test server doesn't have a
`publisher` field in that record, so the function failed. The new code
falls back to using the host name as the server name if it can't get
a `publisher` field value.